PR feedback and move aggregation sig related functions into `aggregation.py`
