+++ tests/beacon/test_aggregation.py
@@ -1,0 +1,68 @@
+import pytest
+from hypothesis import (
+    given,
+    settings,
+    strategies as st,
+)
+
+from eth.utils import bls
+from eth.utils.bitfield import (
+    get_empty_bitfield,
+    has_voted,
+)
+from eth.beacon.aggregation import (
+    aggregate_votes,
+    verify_votes,
+)
+
+
+@settings(max_examples=1)
+@given(random=st.randoms())
+@pytest.mark.parametrize(
+    (
+        'votes_count'
+    ),
+    [
+        (0),
+        (9),
+    ],
+)
+def test_aggregate_votes(votes_count, random, privkeys, pubkeys):
+    bit_count = 10
+    pre_bitfield = get_empty_bitfield(bit_count)
+    pre_sigs = ()
+
+    random_votes = random.sample([i for i in range(bit_count)], votes_count)
+    message = b'hello'
+
+    # Get votes: (committee_index, sig, public_key)
+    votes = [
+        (committee_index, bls.sign(message, privkeys[committee_index]), pubkeys[committee_index])
+        for committee_index in random_votes
+    ]
+
+    # Verify
+    sigs, committee_indices = verify_votes(message, votes)
+
+    # Aggregate the votes
+    bitfield, sigs = aggregate_votes(
+        bitfield=pre_bitfield,
+        sigs=pre_sigs,
+        voting_sigs=sigs,
+        voting_committee_indices=committee_indices
+    )
+
+    try:
+        _, _, pubs = zip(*votes)
+    except ValueError:
+        pubs = ()
+
+    voted_index = [
+        committee_index
+        for committee_index in random_votes
+        if has_voted(bitfield, committee_index)
+    ]
+    assert len(voted_index) == len(votes)
+
+    aggregated_pubs = bls.aggregate_pubs(pubs)
+    assert bls.verify(message, aggregated_pubs, sigs)
