+++ tests/beacon/test_helpers.py
@@ -14,7 +14,6 @@ from eth.beacon.types.attestation_records import AttestationRecord
 from eth.beacon.types.shard_and_committees import ShardAndCommittee
 from eth.beacon.helpers import (
     _get_element_from_recent_list,
-    aggregate_votes,
     get_attestation_indices,
     get_block_hash,
     get_hashes_from_recent_block_hashes,
@@ -23,7 +22,6 @@ from eth.beacon.helpers import (
     get_shards_and_committees_for_slot,
     get_signed_parent_hashes,
     get_block_committees_info,
-    verify_votes,
 )
 
 from tests.beacon.helpers import (
@@ -400,55 +398,3 @@ def test_get_block_committees_info(monkeypatch,
             block_committees_info.proposer_index_in_committee ==
             result_proposer_index_in_committee
         )
-
-
-@settings(max_examples=1)
-@given(random=st.randoms())
-@pytest.mark.parametrize(
-    (
-        'votes_count'
-    ),
-    [
-        (0),
-        (9),
-    ],
-)
-def test_aggregate_votes(votes_count, random, privkeys, pubkeys):
-    bit_count = 10
-    pre_bitfield = get_empty_bitfield(bit_count)
-    pre_sigs = ()
-
-    random_votes = random.sample([i for i in range(bit_count)], votes_count)
-    message = b'hello'
-
-    # Get votes: (committee_index, sig, public_key)
-    votes = [
-        (committee_index, bls.sign(message, privkeys[committee_index]), pubkeys[committee_index])
-        for committee_index in random_votes
-    ]
-
-    # Verify
-    sigs, committee_indices = verify_votes(message, votes)
-
-    # Aggregate the votes
-    bitfield, sigs = aggregate_votes(
-        bitfield=pre_bitfield,
-        sigs=pre_sigs,
-        voting_sigs=sigs,
-        voting_committee_indices=committee_indices
-    )
-
-    try:
-        _, _, pubs = zip(*votes)
-    except ValueError:
-        pubs = ()
-
-    voted_index = [
-        committee_index
-        for committee_index in random_votes
-        if has_voted(bitfield, committee_index)
-    ]
-    assert len(voted_index) == len(votes)
-
-    aggregated_pubs = bls.aggregate_pubs(pubs)
-    assert bls.verify(message, aggregated_pubs, sigs)
